+++ asreview/webapp/src/ProjectComponents/SetupComponents/DataUploadComponents/DatasetFromFile.js
@@ -70,6 +70,7 @@ const rejectStyle = {
 };
 
 const DatasetFromFile = ({ project_id, mode, setDataset }) => {
+
   const {
     error: createProjectError,
     isError: isCreatingProjectError,
