+++ asreview/webapp/src/ProjectComponents/index.js
@@ -1,4 +1,3 @@
-export { default as ImportFromFile } from "./ImportFromFile";
 export { default as ImportProject } from "./ImportProject";
 export { default as ProjectDeleteDialog } from "./ProjectDeleteDialog";
 export { default as ProjectPage } from "./ProjectPage";
